requirements.txt
django~=4.1
pandas~=2.1.1
djangorestframework~=3.14.0